# src/main/resources/com/sonicle/webtop/mail/openapi-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "v1",
    "title": "WebTop Mail",
    "description": "## Request header: X-WT-Mail-Machinery\nAny `/me/*` request may include the header `X-WT-Mail-Machinery: full` to declare that the caller (e.g. a mobile mail app) needs the user's full mail machinery (accounts, folder caches, IMAP idle) running server-side. With it, folder listings and unread counts are served from the live folder cache (the same values the WebTop web client shows) and message listings ride the warm cached paths. The first call bearing the header may take a few seconds while the machinery starts; subsequent calls are fast. The machinery is torn down automatically after ~15 minutes without any call for that user; the next call bearing the header transparently restarts it, so clients should simply send the header on every request and need no eviction handling. Callers that do not need live data (bulk integrations, occasional automation) should omit the header: they are served through per-call paths and never spin up per-user machinery.\n\n---\nGenerator notes:\nhttps://github.com/OpenAPITools/openapi-generator/issues/10010\nhttps://github.com/stoplightio/prism/discussions/523\nhttps://github.com/OpenAPITools/openapi-generator/issues/3172",
    "contact": {
      "email": "dev-team@sonicle.com"
    }
  },
...
  "security": [
    {
      "basicAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/me/messages/list": {
      "parameters": [],
      "get": {
        "summary": "Lists messages envelops of folder",
        "operationId": "listMessages",
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/folderId"
          },
          {
            "$ref": "#/components/parameters/pageNo"
          },
          {
            "$ref": "#/components/parameters/pageSize"
          },
          {
            "$ref": "#/components/parameters/filter"
          },
          {
            "$ref": "#/components/parameters/orderBy"
          },
          {
            "$ref": "#/components/parameters/returnCount"
          }
        ],
        "tags": [
          "me_messages"
        ]
      }
    },
    "/me/messages/send": {
      "parameters": [],
      "post": {
        "summary": "Send new message",
        "operationId": "sendMessage",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageNew"
              }
            }
          }
        },
        "tags": [
          "me_messages"
        ],
        "description": ""
      }
    },
    "/me/messages": {
      "parameters": [],
      "get": {
        "summary": "Get full message of folder by uid",
        "operationId": "getMessage",
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/folderId"
          },
          {
            "$ref": "#/components/parameters/uid"
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "set_seen",
            "description": "Set message as seen"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "index",
            "description": "Attachment index for inner eml"
          }
        ],
        "tags": [
          "me_messages"
        ]
      },
      "delete": {
        "summary": "Trash message by folder and uid",
        "operationId": "deleteMessage",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "folder_id"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "uid"
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "trash"
          }
        ],
        "tags": [
          "me_messages"
        ]
      },
      "put": {
        "summary": "",
        "operationId": "moveMessage",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Move message to another folder",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "from_folder_id"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "to_folder_id"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "uid"
          }
        ],
        "tags": [
          "me_messages"
        ]
      }
    },
    "/me/messages/reply": {
      "parameters": [],
      "get": {
        "summary": "Get reply message of folder by uid",
        "operationId": "getReplyMessage",
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageNew"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/folderId"
          },
          {
            "$ref": "#/components/parameters/uid"
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "replyAll"
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "includeAttachments"
          }
        ],
        "tags": [
          "me_messages"
        ]
      }
    },
    "/me/messages/forward": {
      "parameters": [],
      "get": {
        "summary": "Get forward message of folder by uid",
        "operationId": "getForwardMessage",
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageNew"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/folderId"
          },
          {
            "$ref": "#/components/parameters/uid"
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "includeAttachments"
          }
        ],
        "tags": [
          "me_messages"
        ]
      }
    },
    "/me/messages/attachments/$value": {
      "parameters": [],
      "get": {
        "summary": "Get attachment bytes from message of folder by uid / index",
        "operationId": "getMessageAttachmentBytes",
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/folderId"
          },
          {
            "$ref": "#/components/parameters/uid"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "index",
            "description": "The index of the attachment as received during get"
          }
        ],
        "tags": [
          "me_messages"
        ]
      }
    },
    "/me/messages/cids/$value": {
      "parameters": [],
      "get": {
        "summary": "Get Cid bytes from message of folder by uid / cidName",
        "operationId": "getMessageCidBytes",
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/folderId"
          },
          {
            "$ref": "#/components/parameters/uid"
          },
          {
            "$ref": "#/components/parameters/cidName"
          }
        ],
        "tags": [
          "me_messages"
        ]
      }
    },
    "/me/messages/notes": {
      "parameters": [],
      "get": {
        "summary": "Get Note from message",
        "operationId": "getMessageNote",
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/folderId"
          },
          {
            "$ref": "#/components/parameters/uid"
          }
        ],
        "tags": [
          "me_messages"
        ]
      },
      "post": {
        "summary": "Set Message Note Text",
        "operationId": "setMessageNote",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "me_messages"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Note"
              }
            }
          }
        }
      }
    },
    "/me/messages/seen": {
      "parameters": [],
      "get": {
        "summary": "Get message seen state by folder and uid",
        "operationId": "getMessageSeenState",
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/folderId"
          },
          {
            "$ref": "#/components/parameters/uid"
          }
        ],
        "tags": [
          "me_messages"
        ]
      },
      "post": {
        "summary": "Set Message Seen State",
        "operationId": "setMessageSeenState",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "me_messages"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "folder_id"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "uid"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "boolean"
              }
            }
          }
        }
      }
    },
    "/me/messages/tags": {
      "parameters": [],
      "get": {
        "summary": "Get message tags by folder and uid",
        "operationId": "getMessageTags",
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "x-stoplight": {
                      "id": "hab9osygdik2w"
                    },
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/folderId"
          },
          {
            "$ref": "#/components/parameters/uid"
          }
        ],
        "tags": [
          "me_messages"
        ]
      },
      "post": {
        "summary": "Set Message Tags",
        "operationId": "setMessageTags",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "me_messages"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "folder_id"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "uid"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "x-stoplight": {
                    "id": "olb4u46ieeg4z"
                  },
                  "type": "string"
                }
              }
            }
          }
        },
        "description": ""
      }
    },
    "/me/messages/flags": {
      "parameters": [],
      "get": {
        "summary": "Get message flag by folder and uid",
        "operationId": "getMessageFlag",
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/folderId"
          },
          {
            "$ref": "#/components/parameters/uid"
          }
        ],
        "tags": [
          "me_messages"
        ]
      },
      "post": {
        "summary": "Set Message Flag",
        "operationId": "setMessageFlag",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "me_messages"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "folder_id"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "uid"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "description": ""
      }
    },
    "/me/messages/calendar/apply": {
      "post": {
        "summary": "Apply an iTIP action to a calendar attachment",
        "operationId": "applyCalendarPart",
        "description": "Crosses the caller-supplied action with the iTIP METHOD on the referenced text/calendar attachment and performs the appropriate calendar mutation (create/update/delete) plus, when notify=true, emits an iTIP REPLY mail to the organizer. Single entry point for Accept / Tentative / Decline of REQUEST; Apply update on UPDATE; Remove on CANCEL; Apply attendee response on REPLY; Import on PUBLISH.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItipApplyResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItipApplyRequest"
              }
            }
          }
        },
        "tags": [
          "me_messages"
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "uid": {
            "type": "integer"
          },
          "subject": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "sender": {
            "$ref": "#/components/schemas/Contact"
          },
          "recipients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contact"
            }
          },
          "cc": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contact"
            }
          },
          "bcc": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contact"
            }
          },
          "date": {
            "type": "string"
          },
          "isRead": {
            "type": "boolean"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "status": {
            "type": "string"
          },
          "flag": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hasNote": {
            "type": "boolean",
            "x-stoplight": {
              "id": "ehuxy7g25h8t9"
            }
          },
          "calendarParts": {
            "description": "Populated when one or more text/calendar attachments are present. Empty or omitted otherwise. Each entry corresponds to one ICS attachment carried by the message.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalendarPart"
            }
          }
        }
      },
...
    },
    "parameters": {
      "folderId": {
        "name": "folder_id",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "The full folder name"
      },
      "pageNo": {
        "name": "page_no",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer"
        },
        "description": "The page number to return, providing a value actually activates pagination. Optional."
      },
      "pageSize": {
        "name": "page_size",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer"
        },
        "description": "How many items to return when paginating. Defaults to 50."
      },
      "uid": {
        "name": "uid",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "The message UID in a folder"
      },
      "cidName": {
        "name": "cid_name",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "The Content-ID name of an inline attachment"
      },
      "filter": {
        "name": "filter",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "A RSQL filter query to filter out resulting items. Optional."
      },
      "orderBy": {
        "name": "orderBy",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Field namee and direction (ASC or DESC) to sort resulting items. Optional."
      },
      "returnCount": {
        "name": "returnCount",
        "in": "query",
        "required": false,
        "schema": {
          "type": "boolean"
        },
        "description": "Specifies whether to compute and return the full count of a list of items. Useful when dealing with paginated data. Optional."
      }
    }
  }
}